% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electedBR-package.R
\docType{package}
\name{electedBR-package}
\alias{electedBR}
\alias{electedBR-package}
\title{electedBR: Brazilian Election Winners and Sitting Members of Congress}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Retrieves the candidates elected in Brazilian municipal and general elections since 2018, consolidated from the open data of the Superior Electoral Court (TSE, \url{https://dadosabertos.tse.jus.br/}) and distributed as yearly Parquet files, together with the federal deputies and senators currently serving according to the open data APIs of the Chamber of Deputies (\url{https://dadosabertos.camara.leg.br/}) and the Federal Senate (\url{https://www12.senado.leg.br/dados-abertos}). Election results and current office holding are kept as distinct queries; electoral roles, service history and provenance are preserved. Every function returns a tibble with English column names and has a Portuguese alias.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/electedBR}
  \item \url{https://strategicprojects.github.io/electedBR/}
  \item Report bugs at \url{https://github.com/StrategicProjects/electedBR/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})

Authors:
\itemize{
  \item Andre Leite \email{leite@castlab.org} (\href{https://orcid.org/0000-0002-4718-9766}{ORCID})
  \item Marcos Wasiliew \email{marcos.wasiliew@gmail.com} (\href{https://orcid.org/0009-0004-4694-3159}{ORCID})
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br} (\href{https://orcid.org/0000-0001-6249-0920}{ORCID})
  \item Carlos Amorim \email{carlos.agaf@ufpe.br} (\href{https://orcid.org/0000-0001-6315-8305}{ORCID})
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br} (\href{https://orcid.org/0000-0002-1216-8674}{ORCID})
  \item Júlia Nascimento Barreto \email{juliabarreto@gd.seplag.pe.gov.br} (\href{https://orcid.org/0009-0004-2851-7770}{ORCID})
}

}
\keyword{internal}
